[ACCOUNT] Split pages (#826)
* split pages

* remove WrapperWithNav

* move route to Account/index

- make routes within
